requirements.txt
streamlit>=1.30.0
plotly>=5.18.0
pandas>=2.1.0
numpy>=1.26.0
fpdf2>=2.7.0
openpyxl>=3.1.0
scipy>=1.11.0
